+++ meta/FlexibleSUSY.m
@@ -62,7 +62,8 @@ BeginPackage["FlexibleSUSY`",
               "WeinbergAngle`",
               "Wrappers`",
               "Himalaya`",
-              "GM2Calc`"
+              "GM2Calc`",
+              "Unitarity`"
 }];
 
 $flexiblesusyMetaDir     = DirectoryName[FindFile[$Input]];
@@ -214,6 +215,7 @@ IMEXTPAR = {};
 FSCalculateDecays = False;
 FSDecayParticles = Automatic;
 FSEnableParallelism = True;
+FSUnitarityConstraints = True;
 FSEnableCompile;
 
 (* Standard Model input parameters (SLHA input parameters) *)
@@ -423,6 +425,7 @@ ReplaceSymbolsInUserInput[rules_] :=
            IMEXTPAR                              = IMEXTPAR                              /. rules;
            FlexibleSUSY`ExtraSLHAOutputBlocks    = FlexibleSUSY`ExtraSLHAOutputBlocks    /. rules;
            FlexibleSUSY`FSCalculateDecays        = FlexibleSUSY`FSCalculateDecays        /. rules;
+           FlexibleSUSY`FSUnitarityConstraints   = FlexibleSUSY`FSUnitarityConstraints   /. rules;
            FlexibleSUSY`FSDecayParticles         = FlexibleSUSY`FSDecayParticles         /. rules;
            FlexibleSUSY`FSExtraInputParameters   = FlexibleSUSY`FSExtraInputParameters   /. rules;
            FlexibleSUSY`FSAuxiliaryParameterInfo = FlexibleSUSY`FSAuxiliaryParameterInfo /. rules;
@@ -497,6 +500,15 @@ ReplaceSymbolsInUserInput[rules_] :=
                  " Disabling decays."
               ];
               FlexibleSUSY`FSCalculateDecays = False
+           ];
+
+           With[{sarahVersion = DecomposeVersionString[SA`Version],
+                 minimRequired = {4, 13, 0}},
+              If[!TrueQ@Utils`VersionOrderGtEqThan[sarahVersion, minimRequired],
+                 Print["Warning: using SARAH version ", SA`Version, " but unitarity calculation requires ", StringRiffle[ToString/@minimRequired, "."]];
+                 Print["         Disabling unitarity calculation."];
+                 FlexibleSUSY`FSUnitarityConstraints = False;
+              ];
            ];
           ];
 
@@ -2515,6 +2527,16 @@ WriteFToFConversionInNucleusClass[leptonPairs:{{_->_,_}...}, files_List] :=
       DeleteDuplicates@Join[vertices,npfVertices]
    ];
 
+WriteUnitarityClass[files_List] :=
+    Module[{res},
+      res = If[FSUnitarityConstraints, Unitarity`GetScatteringMatrix[], {0, ""}];
+      WriteOut`ReplaceInFiles[files,
+        {"@scatteringPairsLength@" -> ToString@res[[1]],
+         "@infiniteSMatrix@" -> res[[2]],
+         Sequence @@ GeneralReplacementRules[]
+        }];
+    ];
+
 (* Write the AMM c++ files *)
 WriteAMMClass[fields_List, files_List] :=
     Module[{calculation, getMLCP,
@@ -2876,6 +2898,11 @@ WriteUserExample[inputParameters_List, files_List] :=
                             "@writeCmdLineOutput@" -> IndentText[writeCmdLineOutput],
                             "@fillSLHAIO@" -> fillSLHAIO,
                             "@decaySetttingsOverride@" -> IndentText[decaySetttingsOverride],
+                            "@calculateUnitarity@" -> If[FSUnitarityConstraints,
+                                 "const auto unitarityStruct = " <> FSModelName <> "_unitarity::max_scattering_eigenvalue_infinite_s(std::get<0>(models));\n" <>
+                                  "slha_io.set_unitarity_infinite_s(spectrum_generator_settings, unitarityStruct);",
+                                  ""
+                               ],
                             Sequence @@ GeneralReplacementRules[]
                           } ];
           ];
@@ -2914,6 +2941,7 @@ WriteMathLink[inputParameters_List, extraSLHAOutputBlocks_List, files_List] :=
             calculateModelDecaysFunction = "", fillDecaysSLHA = "", getDecaysVirtualFunc = "",
             getSpectrumDecays = "", putDecaysPrototype = "", putDecaysFunction = "",
             mathlinkDecaysCalculationFunction = "", loadCalculateDecaysFunction = "",
+            setUnitarity = "", loadCalculateUnitarityFunction = "", calculateUnitarityMessages = "",
             calculateDecaysMessages = "", calculateDecaysExample = "", decaysIncludes = "", fdDefaultSettings = "",
             addFDOptions1 = "", addFDOptions2 = "", setFDOptions = "", setDecayOptions = "", fillFDSettings = "",
             decayIndex = "const Index_t n_fd_settings = 0;"},
@@ -2939,6 +2967,14 @@ WriteMathLink[inputParameters_List, extraSLHAOutputBlocks_List, files_List] :=
               defaultSolverType = "-1",
               defaultSolverType = GetBVPSolverSLHAOptionKey[FlexibleSUSY`FSBVPSolvers[[1]]];
              ];
+           If[FSUnitarityConstraints,
+              loadCalculateUnitarityFunction = "FS" <> FlexibleSUSY`FSModelName <> "CalculateUnitarity = LibraryFunctionLoad[lib" <>
+                                            FlexibleSUSY`FSModelName <> ", \"FS" <> FlexibleSUSY`FSModelName <>
+                                            "CalculateUnitarity\", LinkObject, LinkObject];\n";
+              calculateUnitarityMessages = "\n" <> "FS" <> FlexibleSUSY`FSModelName <> "CalculateUnitarity::error = \"`1`\";\n" <>
+                                        "FS" <> FlexibleSUSY`FSModelName <> "CalculateUnitarity::warning = \"`1`\";\n";
+              setUnitarity = "slha_io.set_unitarity_infinite_s(settings, unitarityData);";
+           ];
            If[FlexibleSUSY`FSCalculateDecays,
               decaysData = FlexibleSUSY`FSModelName <> "_decays decays{};              ///< decays";
               getDecaysVirtualFunc = FSMathLink`CreateSpectrumDecaysGetterInterface[FlexibleSUSY`FSModelName];
@@ -3033,6 +3069,9 @@ fillFDSettings = "data.set_fd_settings(flexibledecay_settings);\n"
                             "@setDecayOptions@" -> IndentText @ setDecayOptions,
                             "@fillFDSettings@" -> fillFDSettings,
                             "@decayIndex@" -> decayIndex,
+                            "@loadCalculateUnitarityFunction@" -> loadCalculateUnitarityFunction,
+                            "@calculateUnitarityMessages@" -> calculateUnitarityMessages,
+                            "@setUnitarity@" -> setUnitarity,
                             Sequence @@ GeneralReplacementRules[]
                           } ];
           ];
@@ -3108,7 +3147,8 @@ WriteUtilitiesClass[massMatrices_List, betaFun_List, inputParameters_List, extra
             numberOfDRbarBlocks, drBarBlockNames,
             setDecaysPrototypes = "", setDecaysFunctions = "",
             fillDecaysDataPrototypes = "", fillDecaysDataFunctions = "",
-            decaysHeaderIncludes = "", useDecaysData = ""
+            decaysHeaderIncludes = "", useDecaysData = "",
+            unitarityIncludes = ""
            },
            particles = DeleteDuplicates @ Flatten[TreeMasses`GetMassEigenstate /@ massMatrices];
            susyParticles = Select[particles, (!TreeMasses`IsSMParticle[#])&];
@@ -3165,6 +3205,9 @@ WriteUtilitiesClass[massMatrices_List, betaFun_List, inputParameters_List, extra
            drBarBlockNames      = WriteOut`GetDRbarBlockNames[lesHouchesParameters];
            gaugeCouplingNormalizationDecls = WriteOut`GetGaugeCouplingNormalizationsDecls[SARAH`Gauge];
            gaugeCouplingNormalizationDefs  = WriteOut`GetGaugeCouplingNormalizationsDefs[SARAH`Gauge];
+           If[FSUnitarityConstraints,
+              unitarityIncludes = "#include \"" <> FlexibleSUSY`FSModelName <> "_unitarity.hpp\""
+           ];
            If[FlexibleSUSY`FSCalculateDecays,
               setDecaysPrototypes = WriteOut`CreateSetDecaysPrototypes[FlexibleSUSY`FSModelName];
               setDecaysFunctions = WriteOut`CreateSetDecaysFunctions[FlexibleSUSY`FSModelName];
@@ -3228,6 +3271,7 @@ WriteUtilitiesClass[massMatrices_List, betaFun_List, inputParameters_List, extra
                             "@useDecaysData@"                   -> useDecaysData,
                             "@numberOfNeutralGoldstones@"       -> IndentText["static constexpr int number_of_neutral_goldstones = " <> ToString[TreeMasses`GetDimensionStartSkippingGoldstones[TreeMasses`GetPseudoscalarHiggsBoson[]]-1] <> ";"],
                             "@numberOfChargedGoldstones@"       -> IndentText["static constexpr int number_of_charged_goldstones = " <> ToString[TreeMasses`GetDimensionStartSkippingGoldstones[TreeMasses`GetChargedHiggsBoson[]]-1] <> ";"],
+                            "@unitarityIncludes@" -> unitarityIncludes,
                             Sequence @@ GeneralReplacementRules[]
                           } ];
           ];
@@ -5277,6 +5321,13 @@ MakeFlexibleSUSY[OptionsPattern[]] :=
                      {FileNameJoin[{$flexiblesusyTemplateDir, "FFV_form_factors.cpp.in"}],
                              FileNameJoin[{FSOutputDir, FlexibleSUSY`FSModelName <> "_FFV_form_factors.cpp"}]}}
                ];
+
+           Print["Creating unitarity class..."];
+           WriteUnitarityClass[{{FileNameJoin[{$flexiblesusyTemplateDir, "unitarity.hpp.in"}],
+                               FileNameJoin[{FSOutputDir, FlexibleSUSY`FSModelName <> "_unitarity.hpp"}]},
+                           {FileNameJoin[{$flexiblesusyTemplateDir, "unitarity.cpp.in"}],
+                               FileNameJoin[{FSOutputDir, FlexibleSUSY`FSModelName <> "_unitarity.cpp"}]}}
+           ];
 
            Print["Creating C++ QFT class..."];
            cxxQFTTemplateDir = FileNameJoin[{$flexiblesusyTemplateDir, "cxx_qft"}];