+++ src/slha_io.hpp
@@ -21,6 +21,7 @@
 
 #include "slha_format.hpp"
 #include "decays/decay.hpp"
+#include "unitarity.hpp"
 
 #include <complex>
 #include <functional>
@@ -154,6 +155,7 @@ public:
    void set_settings(const Spectrum_generator_settings&);
    void set_FlexibleDecay_settings(const FlexibleDecay_settings&);
    void set_sminputs(const softsusy::QedQcd&);
+   void set_unitarity_infinite_s(Spectrum_generator_settings const&, UnitarityInfiniteS const&);
    void write_to_file(const std::string&) const;
    void write_to_stream() const;
    void write_to_stream(std::ostream&) const;